Maintnance (#33)
* Fix OGL framebuffer vertically fliped issue

* Move editor font for imgui to editor project

* Fix passing log arguments to assert macro

* Clean up unused vulkan files

* Clean up TODOs

* WIP and change to integers for colors and texIndex

* Move vsync setting to graphics context

+++ assets/shaders/basic_vertex.glsl
@@ -9,11 +9,11 @@
 layout(location = 0) in vec2 a_Position;
 layout(location = 1) in vec2 a_TexCoord;
 layout(location = 2) in vec4 a_Color;
-layout(location = 3) in float a_TexIndex;
+layout(location = 3) in uint a_TexIndex;
 
 layout(location = 0) out vec2 v_TexCoord;
 layout(location = 1) out vec4 v_Color;
-layout(location = 2) out float v_TexIndex;
+layout(location = 2) out uint v_TexIndex;
 
 layout(std140, binding = 0) uniform Camera
 {
